requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0 